+++ src/main/java/com/caoccao/javet/interop/IV8Creatable.java
@@ -43,6 +43,9 @@ public interface IV8Creatable {
     @CheckReturnValue
     V8ValueFunction createV8ValueFunction(JavetCallbackContext javetCallbackContext) throws JavetException;
 
+    @CheckReturnValue
+    V8ValueFunction createV8ValueFunction(String codeString) throws JavetException;
+
     V8ValueInteger createV8ValueInteger(int integerValue) throws JavetException;
 
     V8ValueLong createV8ValueLong(long longValue) throws JavetException;