+++ src/test/java/com/caoccao/javet/interop/converters/TestJavetProxyConverter.java
@@ -46,13 +46,12 @@ public class TestJavetProxyConverter extends BaseTestJavetRuntime {
         v8Runtime.setConverter(javetProxyConverter);
     }
 
-    // TODO
-//    @Test
+    @Test
     public void testConstructor() throws JavetException {
         v8Runtime.getGlobalObject().set("StringBuilder", StringBuilder.class);
         assertEquals("abc def", v8Runtime.getExecutor(
                 "function main() {\n" +
-                        "  return new StringBuilder().append('abc').append(' ').append('def').toString();" +
+                        "  return new StringBuilder().append('abc').append(' ').append('def').toString();\n" +
                         "}\n" +
                         "main();").executeString());
         v8Runtime.getGlobalObject().delete("StringBuilder");