+++ docs/release_notes.rst
@@ -7,6 +7,7 @@ Release Notes
 
 * Added ``resolve()`` and ``reject()`` to ``V8ValuePromise``
 * Added ``staticClassEnabled`` to ``JavetProxyConverter``
+* Added ``construct()`` to ``JavetUniversalProxyHandler``
 * Added Dockerfile for Linux x86-64
 * Refactored document for how to build Javet